Removes Advocates store because it's not needed
